refactor(oauth-as): extract src/lib/oauth into the module root (KAN-415)
The pilot extraction. `modules.json` marks `oauth-as` riskTier=critical with
five frozen contracts (URLs, iss derivation, scope set, RS256/JWKS keys, exact
401 shape) and states plainly: "claude.ai, Claude Desktop and MCP Inspector are
external consumers. Contract test required BEFORE any move." That test landed
first and is mutation-proven; this commit is the move it was gating.

WHAT MOVED, AND WHAT DELIBERATELY DID NOT
  src/lib/oauth/**  ->  src/modules/oauth-as/lib/**   (10 files, git-detected renames)
  src/app/oauth/**                                     UNCHANGED
  src/app/api/well-known/**                            UNCHANGED

In the App Router a route's URL IS its directory path, so the route files stay
exactly where they are. URLs are frozen by construction, not by care — there is
no edit that could have changed one. The contract test confirms it: 8/8 green
before and after, asserting a live handler still answers every endpoint the
discovery document advertises.

ESTATE REWORK, IN THE SAME PR (KAN-428 DoD)
  .github/CODEOWNERS      /src/lib/oauth/ was about to become a dead entry —
                          i.e. the most audit-sensitive module in the product
                          merging unreviewed. This is the exact failure the DoD
                          gate exists to catch.
  modules.json            oauth-as paths + the serviceRoleClient file list
  env-access-baseline.json  CTL-037's two entries; the guard re-verifies
                          72 reads across 37 files, matching exactly
  kan416-derive.py        the path->module mapping the manifest regenerates from

TEST COUPLING — the F4 thesis, demonstrated
Two tests broke on the move, and neither for a reason about OAuth:

1. dcr-anti-phishing.test.ts died with ENOENT readFileSync'ing a moved file.
   The path came from the manifest, so it was a one-entry fix. That is F4
   working.
2. The same file asserted an import with the literal '@/lib/oauth/client-trust'
   baked into a regex, so a wiring assertion failed for reasons unrelated to
   wiring. The specifier is now DERIVED from the manifest, with a guard on the
   derivation itself so a key that went undefined cannot build a pattern that
   matches nothing while looking green.

Not weakened, not skipped: both still assert exactly what they did before.

150/150 across the 15 oauth suites. 0 real-source tsc errors.

UI-Change-Approved: KAN-415

The only UI-protected file touched is src/app/oauth/authorize/page.tsx, and the
diff is three import specifiers — no design change and no copy change. Luisa
approved this in session on 2026-08-04, on exactly that basis.
