+++ src/app/api/well-known/oauth-authorization-server/route.ts
@@ -10,7 +10,7 @@
  */
 
 import { NextResponse } from 'next/server';
-import { oauthConfig } from '@/lib/oauth/config';
+import { oauthConfig } from '@/modules/oauth-as/lib/config';
 
 export const dynamic = 'force-dynamic'; // env-driven URLs — never cache
