+++ src/modules/oauth-as/lib/config.ts
