+++ tests/support/source-paths.json
@@ -1,7 +1,7 @@
 {
   "$comment": "GENERATED — regenerate with `npm run gen:test-paths`. Mirror of tests/support/source-paths.ts for .cjs and shell consumers (KAN-414 F4).",
   "generated_from": "literals referenced by tracked files under tests/",
-  "count": 176,
+  "count": 177,
   "SRC": {
     "codeowners": ".github/CODEOWNERS",
     "pullRequestTemplate": ".github/PULL_REQUEST_TEMPLATE.md",
@@ -155,7 +155,6 @@
     "fileMagicBytes": "src/lib/file-magic-bytes.ts",
     "inviteText": "src/lib/invite-text.ts",
     "moderationPolicy": "src/lib/moderation-policy.ts",
-    "clients": "src/lib/oauth/clients.ts",
     "pipeline": "src/lib/recommender/v2/pipeline.ts",
     "flags": "src/lib/retention/flags.ts",
     "sweep": "src/lib/retention/sweep.ts",
@@ -163,6 +162,8 @@
     "supabaseService": "src/lib/supabase-service.ts",
     "turnstile": "src/lib/turnstile.ts",
     "middleware": "src/middleware.ts",
+    "clientTrust": "src/modules/oauth-as/lib/client-trust.ts",
+    "clients": "src/modules/oauth-as/lib/clients.ts",
     "database": "src/types/database",
     "index": "src/types/database/index.ts",
     "prod": "src/types/database/prod.ts",
