+++ tests/unit/oauth-authorize-suspension-guard.test.ts
@@ -36,7 +36,7 @@
  */
 import { redirect } from 'next/navigation';
 import { getAccountStanding } from '@/lib/account-status';
-import { validateAuthorizeRequest } from '@/lib/oauth/authorize';
+import { validateAuthorizeRequest } from '@/modules/oauth-as/lib/authorize';
 
 jest.mock('next/navigation', () => ({
   redirect: jest.fn((to: string) => {
@@ -59,13 +59,13 @@ jest.mock('@/lib/account-status', () => ({
   getAccountStanding: jest.fn(),
 }));
 
-jest.mock('@/lib/oauth/authorize', () => ({
+jest.mock('@/modules/oauth-as/lib/authorize', () => ({
   validateAuthorizeRequest: jest.fn(),
   buildErrorRedirect: jest.fn(() => '/oauth/error'),
 }));
 
-jest.mock('@/lib/oauth/consents', () => ({ getConsent: jest.fn(async () => null) }));
-jest.mock('@/lib/oauth/client-trust', () => ({
+jest.mock('@/modules/oauth-as/lib/consents', () => ({ getConsent: jest.fn(async () => null) }));
+jest.mock('@/modules/oauth-as/lib/client-trust', () => ({
   clientTrust: jest.fn(() => ({ level: 'unknown' })),
   redirectHost: jest.fn(() => 'example.test'),
 }));
