+++ tests/unit/oauth/dcr-anti-phishing.test.ts
@@ -16,9 +16,25 @@ describe('consent screen wires the trust surface (SEC-76 web-oauth-7)', () => {
   const src = fs.readFileSync(path.join(ROOT, SRC.authorizePage), 'utf8');
 
   test('imports clientTrust + redirectHost from the helper module', () => {
-    expect(src).toMatch(
-      /import\s+\{[^}]*clientTrust[^}]*redirectHost[^}]*\}\s+from\s+['"]@\/lib\/oauth\/client-trust['"]/
+    // The module SPECIFIER is derived from the manifest, not written out. It
+    // used to be the literal '@/lib/oauth/client-trust', and the `oauth-as`
+    // extraction (KAN-415) broke it — an assertion about wiring failing for a
+    // reason that has nothing to do with wiring, which is the KAN-417 coupling
+    // this manifest exists to remove. Deriving it means the next move updates
+    // one entry instead of this regex.
+    const specifier = `@/${SRC.clientTrust.replace(/^src\//, '').replace(/\.ts$/, '')}`;
+    const importRe = new RegExp(
+      `import\\s+\\{[^}]*clientTrust[^}]*redirectHost[^}]*\\}\\s+from\\s+['"]${specifier.replace(
+        /[.*+?^${}()|[\]\\]/g,
+        '\\$&',
+      )}['"]`,
     );
+
+    // Guard the derivation itself: a manifest key that silently became
+    // undefined would build the regex `@/undefined` and fail confusingly, or —
+    // worse, if the pattern were looser — match nothing while looking green.
+    expect(SRC.clientTrust).toMatch(/client-trust\.ts$/);
+    expect(src).toMatch(importRe);
   });
 
   test('computes trust from the client is_first_party flag', () => {
